Wrap each question/answer pair with bos and eos tokens
This may help with the hallucination issues by providing the model with
clear boundaries between question and answer pairs. Note that we also
use the correct eos token from the model config instead of the one from
the tokenizer. This is due to the fact that OPT-based BLIP2 changed its
eos token from `</s>` to `\n`, and this is reflected in the model
config, but not in the tokenizer.